Move some files about so that we can add more c++ versions
