+++ BUILD.bazel
@@ -18,7 +18,7 @@ cc_test(
 cc_library(
     name = "indirect_cxx11",
     hdrs = [
-        "c++11/indirect_11.h",
+        "compatibility/indirect_cxx11.h",
         "indirect.h",
     ],
     copts = ["-Iexternal/value_types/"],
@@ -56,7 +56,7 @@ cc_test(
 cc_library(
     name = "polymorphic_cxx11",
     hdrs = [
-        "c++11/polymorphic_11.h",
+        "compatibility/polymorphic_cxx11.h",
         "polymorphic.h",
     ],
     defines = ["XYZ_POLYMORPHIC_CXX_11"],
